+++ src/love/config.cpp
@@ -1,0 +1,12 @@
+#include "config.h"
+#include "../ChaiLove.h"
+
+namespace love {
+
+config::config() {
+	options["alphablending"] = true;
+	options["highquality"] = true;
+	version = CHAILOVE_VERSION_STRING;
+}
+
+}  // namespace love