+++ src/love/script.h
@@ -2,7 +2,7 @@
 #define SRC_LOVE_SCRIPT_H_
 
 #include <string>
-#include "Types/System/Config.h"
+#include "config.h"
 
 #ifdef __HAVE_CHAISCRIPT__
 #include <chaiscript/chaiscript.hpp>
@@ -10,8 +10,6 @@
 // TODO(RobLoach): Fix Chai loading with the Standard Library.
 // #include <chaiscript/chaiscript_stdlib.hpp>
 #endif
-
-using love::Types::System::Config;
 
 namespace love {
 
@@ -25,7 +23,7 @@ class script {
 	/**
 	 * @brief Modify some configuration options.
 	 *
-	 * @param t The Config object to modify.
+	 * @param t The config object to modify.
 	 *
 	 * ### Example
 	 *
@@ -36,7 +34,7 @@ class script {
 	 * }
 	 * @endcode
 	 */
-	void conf(Config& t);
+	void conf(config& t);
 
 	/**
 	 * @brief This function is called exactly once at the beginning of the game.
@@ -235,7 +233,7 @@ class script {
 	#ifdef __HAVE_CHAISCRIPT__
 	chaiscript::ChaiScript chai;
 	std::function<void()> chaiload;
-	std::function<void(Config&)> chaiconf;
+	std::function<void(config&)> chaiconf;
 	std::function<void(float)> chaiupdate;
 	std::function<void()> chaidraw;
 	std::function<void()> chaireset;